Solved issue with findAccountsByName() in AccountRepository is returning an empty array.
Issue was due to the query expecting the exact name, including space (ie. "John Doe") which required it to be passed into the parameter with the correct URL encoding for the space character "%20"
eg:
$ curl http://localhost:8080/accounts?name=John%20Doe

Instead implemented a search function findAccountsByNameContainingIgnoreCase() to return all Accounts with a name matching the string. It is closer to the book's examples.

Not related to specific database SQL method (kept using MariaDB for practice).